Removed manual load
