+++ EZMAIL/Socket.php
@@ -3,15 +3,6 @@
 namespace EZMAIL;
 use Exception;
 
-interface ISocket
-{
-    public function open(string $host, int $port, float $timeout) : void;
-    public function readString(int $lenToRead) : string;
-    public function writeString(string $data) : void;
-    public function enableCrypto() : void;
-    public function close() : void;
-}
-
 class Socket implements ISocket
 {
     private $connection; // resource.
