+++ phpunit.xml
@@ -1,11 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<phpunit 
-    bootstrap="tests/autoload.php"
-    colors="true"
-    stopOnFailure="false">
+<phpunit bootstrap="./vendor/autoload.php" colors="true">
     <testsuites>
-        <testsuite name="SMTP">
-            <directory>tests</directory>
+        <testsuite name="Tests">
+            <directory>./tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
