--- tests/autoload.php
@@ -1,8 +1,0 @@
-<?php
-
-require __DIR__ . "/../vendor/autoload.php";
-require __DIR__ . "/../EZMAIL/Socket.php";
-require __DIR__ . "/../EZMAIL/SMTP.php";
-require __DIR__ . "/FakeSocket.php";
-
-?>
